third_party/WebKit/LayoutTests/inspector/profiler/cpu-profiler-profiling-expected.txt

Tests that CPU profiling works. Bug 52634.

Profiler was enabled.

Running: testProfiling
found pageFunction cpu-profiler-profiling.html:7
found (anonymous) 

Profiler was disabled.
